Gigantic overhaul over flexibility
When trying to expand the game to include the second game to much double code was involved. To fix this i started using interfaces and abstract classes inspired by the template and strategy pattenrs.

ElektroMap, ElektroTile, CalcBoardMap and CalcBoardTile are unity objects (monobehaviours) but never use any of the functions belonging to these. Because they are monobehaviour some they are limited (like how they don't like using the construction but initialize). Because they never use these methods i decided to remove them and only use the serializable classes. If these classes need to be moved or filled in a unity object will hold them and do these actions.

Because of this overhaul many of the code is unable to work, currently only the creation of ElektroMap works